# fe/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Backend API",
        "description": "This API exposes endpoints to manage demo.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "security": [
        {
            "Bearer_Authentication": []
        }
    ],
    "paths": {
...
        "/api/v1/daily-meal-plans": {
            "get": {
                "tags": [
                    "daily-meal-plan-controller"
                ],
                "operationId": "getDailyMealPlansByUserAndDate",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2024-07-03"
                        },
                        "example": "2024-07-03"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyMealPlanResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "daily-meal-plan-controller"
                ],
                "operationId": "updateDailyMealPlans",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2024-07-03"
                        },
                        "example": "2024-07-03"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DailyMealPlanDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyMealPlanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "signin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/send-otp": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "generateOTP",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "changePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/checkotp": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "validateOTP",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateOtpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DailyMealPlanDto": {
                "type": "object",
                "properties": {
                    "breakfast": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MealDto"
                        }
                    },
                    "lunch": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MealDto"
                        }
                    },
                    "dinner": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MealDto"
                        }
                    },
                    "nutrients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Nutrient"
                        }
                    }
                }
            },
            "DailyMealPlanResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DailyMealPlanDto"
                    }
                }
            },
            "MealDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "meal_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "typeFood": {
                        "type": "string"
                    }
                }
            },
            "Nutrient": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "percentOfDailyNeeds": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "SignUpRequest": {
                "required": [
                    "confirmPassword",
                    "email",
                    "fullName",
                    "password",
                    "phone"
                ],
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "maxLength": 10,
                        "minLength": 10,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 2147483647,
                        "minLength": 6,
                        "type": "string"
                    },
                    "confirmPassword": {
                        "type": "string"
                    }
                }
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                }
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "role": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    }
                }
            },
            "ChangePasswordRequest": {
                "required": [
                    "confirmPassword",
                    "email",
                    "otp",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 2147483647,
                        "minLength": 6,
                        "type": "string"
                    },
                    "confirmPassword": {
                        "type": "string"
                    },
                    "otp": {
                        "type": "string"
                    }
                }
            },
            "ValidateOtpRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "otp": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer_Authentication": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}